LPS-30029 Regenerate Services

+++ portlets/polls-portlet/docroot/WEB-INF/src/com/liferay/portlet/polls/service/persistence/PollsChoicePersistenceImpl.java
@@ -320,7 +320,9 @@ public class PollsChoicePersistenceImpl extends BasePersistenceImpl<PollsChoice>
 						pollsChoice.getPrimaryKeyObj());
 			}
 
-			session.delete(pollsChoice);
+			if (pollsChoice != null) {
+				session.delete(pollsChoice);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -329,7 +331,9 @@ public class PollsChoicePersistenceImpl extends BasePersistenceImpl<PollsChoice>
 			closeSession(session);
 		}
 
-		clearCache(pollsChoice);
+		if (pollsChoice != null) {
+			clearCache(pollsChoice);
+		}
 
 		return pollsChoice;
 	}
