+++ portlets/polls-portlet/docroot/WEB-INF/src/com/liferay/portlet/polls/service/persistence/PollsQuestionPersistenceImpl.java
@@ -346,7 +346,9 @@ public class PollsQuestionPersistenceImpl extends BasePersistenceImpl<PollsQuest
 						pollsQuestion.getPrimaryKeyObj());
 			}
 
-			session.delete(pollsQuestion);
+			if (pollsQuestion != null) {
+				session.delete(pollsQuestion);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -355,7 +357,9 @@ public class PollsQuestionPersistenceImpl extends BasePersistenceImpl<PollsQuest
 			closeSession(session);
 		}
 
-		clearCache(pollsQuestion);
+		if (pollsQuestion != null) {
+			clearCache(pollsQuestion);
+		}
 
 		return pollsQuestion;
 	}
